fix(windows): revert three over-reaches; slash form is for values, not access
The first pass applied slash form to three things that are ACCESS, not identity, and windows CI caught all
three -- nine cleared, six regressed.

Join builds a path FOR USE and its result goes to the filesystem: Join(a,b,c) answered "a/b/c" where the
platform wants "a\b\c". Reverted to native, with the distinction from Name/Parent stated at the site --
those answer questions ABOUT a path as a value.

commonAncestor's result becomes the run's confinement Root, so it must be a path the platform can resolve.
A slash-form answer made fsroot panic trying to relate C:\... to /. Reverted to native; its test feeds Unix
absolute paths that mean nothing on Windows, so it moves behind //go:build unix -- the Windows behavior
that matters is covered by TestRegisterLayer_Link, which uses the result as a real root.

The file resource identity stays OS-native for now. Normalizing it at that one site fixed the identity
string and broke four consumers keyed on the native form -- including drift detection, which stopped seeing
a stale file. That is the "along the way" fix #547 exists to prevent; the transform belongs at construction
with the resource owning both forms.
